refactor: meeting/config에서 meeting/constant로 stationList 파일 위치 이동
