+++ frontend/src/features/meeting/lib/formValidation.ts
@@ -1,6 +1,6 @@
 import { ValidationError } from '@shared/types/validationError';
 
-import { STATION_LIST } from '../config/stationList';
+import { STATION_LIST } from '../constant/stationList';
 
 const validateDepartureListMinLength = (
   length: number,
