+++ frontend/src/features/recommendation/components/bottomSheetDetail/BottomSheetDetail.tsx
@@ -2,8 +2,8 @@ import { SelectedLocation } from '@features/recommendation/types/SelectedLocatio
 
 import { StartingPlace } from '@entities/location/types/Location';
 
-import { flex, scroll, typography } from '@shared/styles/default.styled';
 import { numberToCharCode } from '@shared/lib/numberToCharCode';
+import { flex, scroll, typography } from '@shared/styles/default.styled';
 
 import DetailSection from '../detailSection/DetailSection';
 import PlaceCard from '../placeCard/PlaceCard';
